+++ src/main/java/roomescape/reservation/repository/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dao;
+package roomescape.reservation.repository;
 
 import java.time.LocalDate;
 import java.util.List;
