+++ src/main/java/roomescape/reservation/repository/JdbcReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dao;
+package roomescape.reservation.repository;
 
 import java.util.List;
 import java.util.Optional;
