+++ src/main/java/roomescape/reservation/repository/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dao;
+package roomescape.reservation.repository;
 
 import java.time.LocalDate;
 import java.util.List;
