+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -3,8 +3,8 @@ package roomescape.reservation.service;
 import java.util.List;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import roomescape.reservation.dao.ReservationRepository;
-import roomescape.reservation.dao.ReservationTimeRepository;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.dto.request.AvailableTimeRequest;
